ubb-sdk/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ubb-sdk"
version = "2.0.0"
description = "Python SDK for the Usage-Based Billing platform"
requires-python = ">=3.10"
dependencies = ["httpx>=0.28", "PyJWT>=2.8"]

[tool.setuptools.packages.find]
include = ["ubb*"]

[tool.setuptools.package-data]
ubb = ["py.typed"]
